changed monetary amounts (cashflow, balance) format to decimal(10,2)

+++ datamodel.py
@@ -4,7 +4,7 @@ db = SQLAlchemy()
 class Balance(db.Model):
     __tablename__ = 'BALANCE'
     id = db.Column(db.Integer, primary_key = True)
-    amount = db.Column(db.Float(precision=2), nullable=False)
+    amount = db.Column(db.Numeric(10,2), nullable=False)
     date = db.Column(db.Date, nullable=False, unique=True)
 
 class Category(db.Model):
@@ -20,7 +20,7 @@ class Cashflow(db.Model):
     __tablename__ = 'CASHFLOW'
     id = db.Column(db.Integer, primary_key=True)
     description = db.Column(db.String(256), nullable=False)
-    amount = db.Column(db.Float(precision=2), nullable=False)
+    amount = db.Column(db.Numeric(10,2), nullable=False)
     date = db.Column(db.Date, nullable=False)
     booked = db.Column(db.Boolean, nullable=False, default=False)
     category_id = db.Column(db.Integer, db.ForeignKey('CATEGORY.id'), nullable=False)
